Update input parsing to use float with comma decimal
Changed default argument values in several exercises to use comma as the decimal separator for float parsing, ensuring compatibility with locales that expect this format. Updated related README and test-data.json files to reflect new input formats and expected outputs. Also changed function signatures and variable types in Exercise10a and Exercise10d from int to double for consistency with input parsing.

+++ Exercise10a_Between10And20OrEven/10a-Between10And20OrEven.cs
@@ -1,12 +1,12 @@
 using System;
 
-bool Between10And20OrEven(int number)
+bool Between10And20OrEven(double number)
 {
     throw new NotImplementedException();
 }
 
 // Obtener argumentos de línea de comandos
-int number = args.Length > 0 ? int.Parse(args[0]) : 15;
+double number = args.Length > 0 ? double.Parse(args[0]) : 15.0;
 
 // Probar la función
 bool result = Between10And20OrEven(number);
